% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novels_excerpts}
\alias{novels_excerpts}
\title{Excerpts from English novels}
\format{A dataframe with 21 rows and 3 variables: \describe{
  \item{title}{Title} \item{author}{Author} \item{text}{Excerpt of text in
  complete sentences from the first 1,000 chars of the novel.} }}
\source{
Novel excerpts obtained from Project Gutenberg full texts in the
  public domain in the USA. \url{http://gutenberg.org}
}
\usage{
novels_excerpts
}
\description{
A dataset of text from English novels by Jane Austen, George Eliot, and
Elizabeth Gaskell.
}
\keyword{datasets}
