% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylest_fit.R
\name{print.stylest_model}
\alias{print.stylest_model}
\title{Custom print method for stylest_model}
\usage{
\method{print}{stylest_model}(x, ...)
}
\arguments{
\item{x}{`stylest_model` object}

\item{...}{Additional arguments}
}
\value{
Prints summary information about the `stylest_model` object
}
\description{
Custom print method for stylest_model
}
\examples{

data(novels_excerpts)
speaker_mod <- stylest_fit(novels_excerpts$text, novels_excerpts$author)
print(speaker_mod)
 
}
