% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylest.R
\name{stylest}
\alias{stylest}
\title{stylest: A package for estimating textual distinctiveness}
\description{
stylest provides a set of functions for fitting a model of speaker
distinctiveness, including tools for selecting the optimal vocabulary for the
model and predicting the most likely speaker (author) of a new text.
}
