% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.stylest_predict}
\alias{summary.stylest_predict}
\title{Summarise Predictions}
\usage{
\method{summary}{stylest_predict}(object, type = c("predicted", "probs",
  "logprobs"), ...)
}
\arguments{
\item{object}{Predictions from a stylest_model
(a \code{\link{stylest_predict}} object)}

\item{type}{Type of prediction "\code{predicted}" (the default),
"\code{logprobs}" or "\code{probs}".  See Details.}

\item{...}{Other arguments (ignored)}
}
\value{
a summary data frame of predictions
}
\description{
This summary function returns a data frame with documents as
rownames and either the most likely author, or posterior
probabilities over all possible authors, for each document.
}
\details{
If \code{type} is "\code{predicted}",
the columns are "\code{author}", a character vector of predicted author
names, and "\code{prob}", the posterior probability of this author
assignment under the model.
If \code{type} is "\code{logprobs}" columns are named by
possible author and each element is
the log posterior probability assigned to each document and author,
relative to the other possible authors.
If \code{type} is "\code{probs}", each element is the posterior
probability that a document is assigned to each author relative to
the others.  This is constructed from the log posterior
unsing a logistic transformation across rows, a.k.a. 'softmax'.

Note: if \code{newdata} is a quanteda::corpus then quanteda's \code{docnames} may return generic names.  These are currently 
"text1", "text2", etc. so you may want to change them first.
}
\examples{
data(novels_excerpts)
speaker_mod <- stylest_fit(novels_excerpts$text, novels_excerpts$author)
preds <- stylest_predict(speaker_mod, "This is an example text, who wrote it?")
summary(preds)
}
