% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylest_fit.R
\name{fit_term_usage}
\alias{fit_term_usage}
\title{Computes speakers' term usage rates}
\usage{
fit_term_usage(x, speaker, terms, smooth)
}
\arguments{
\item{x}{Text vector. May be a \code{corpus_frame} object}

\item{speaker}{Vector of speaker labels. Should be the same length as
\code{x}}

\item{terms}{Vocabulary for document term matrix}

\item{smooth}{Numeric value used smooth term frequencies}
}
\value{
named list of terms, vector of num tokens uttered by each speaker,
  smoothing value, and (smoothed) term usage rate matrix
}
\description{
Computes speakers' term usage rates
}
